add Checkbook::load_from_file

+++ checkbook.h
@@ -39,7 +39,7 @@ class Checkbook {
         Check checks[SIZE];   // array of all the checks
         size_t used;          // number of checks used
         size_t nextCheckNum;  // number of next available check
-        int balance;          // balance of the account
+        double balance;       // balance of the account
 
         // helper functions
         void clearNewlines(std::istream& ins);
